Move hebbian transpose 2d swapped and hebbian 2d to separate files
