+++ hebb/__init__.py
@@ -1,1 +1,3 @@
 from .hebbian_conv_transpose2d import HebbianConvTranspose2d
+from .hebbian_conv_2d import HebbianConv2d
+from .hebbian_conv_transpose2d_swapped import HebbianConvTranspose2dSwapped
